Zjazd6/1_gr2_Michal/fiszki/requirements.txt
django
django-crispy-forms
crispy-bootstrap4
djlint